+++ CHANGELOG.md
@@ -1,7 +1,9 @@
-## Release 0.19.0
+## Release 0.19.0 (2023-06-15)
 * Add Jinja2 template rendering support to the `ksconf_package` module.
 * The `ksconf_package` module is now an "action" which means it has to be run from the controller.
   This allows access to all variables, which is not possible when run from the remote machine, as modules are.
+* Add some validation for `.conf.j2` files; ensuring that they render to valid `.conf` files, and therefore avoid building broken files.
+* Increase atomic file operations (mostly in the ksconf package; when using v0.11.5 or higher)
 
 ## Release 0.18.1 (2023-06-08)
 * NO changes.  Re-uploading the same codebase to workaround issue with Ansible Galaxy where it apparently gets confused about beta releases.